requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
coverage==7.3.0
flake8==6.1.0
idna==3.4
mccabe==0.7.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
pyTelegramBotAPI==4.13.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
six==1.16.0
urllib3==2.0.4
